backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
PyJWT==2.8.0
python-socketio==5.8.0
mysql-connector-python==8.1.0
Werkzeug==2.3.7
 